requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bleach==6.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
kaggle==1.7.4.5
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.6
psutil==7.2.2
pydantic==2.13.1
pydantic_core==2.46.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
scipy==1.13.1
six==1.17.0
starlette==0.49.3
sympy==1.14.0
text-unidecode==1.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.33
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.39.0
webencodings==0.5.1
zipp==3.23.0
